requirements.txt
numpy==1.23.5
opencv_python==4.6.0.66
Pillow==9.1.0
Pillow==10.2.0
Requests==2.31.0
retina_face==0.0.13
skimage==0.0
torch==1.13.1
tqdm==4.65.0
